requirements.txt
Django==6.0
djangorestframework
gunicorn
dj-database-url
psycopg2-binary
whitenoise
requests
python-dateutil
django-humanize